% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lm_topic_labels.R
\name{as.lm_topic_labels}
\alias{as.lm_topic_labels}
\alias{is.lm_topic_labels}
\title{lm_topic_labels Object}
\usage{
as.lm_topic_labels(
  x,
  terms,
  prompts,
  model,
  params,
  with_token,
  time,
  model_output,
  labels
)

is.lm_topic_labels(obj, verbose = FALSE)
}
\arguments{
\item{x}{[\code{named list}]\cr
\code{\link{RollingLDA}} object. Alternatively each element can be passed for
individual results. Individually set elements overwrite elements from \code{x}.}

\item{terms}{[\code{list(n) of character}]\cr
TBA}

\item{prompts}{[\code{character(n)}]\cr
TBA}

\item{model}{[\code{character(1)}]\cr
TBA}

\item{params}{[\code{named list}]\cr
TBA}

\item{with_token}{[\code{logical(1)}]\cr
TBA}

\item{time}{[\code{numeric(1)}]\cr
TBA}

\item{model_output}{[\code{character(n)}]\cr
TBA}

\item{labels}{[\code{character(n)}]\cr
TBA}

\item{obj}{[\code{R} object]\cr
Object to test.}

\item{verbose}{[\code{logical(1)}]\cr
Should test information be given in the console?}
}
\value{
[\code{named list}] \code{\link{lm_topic_labels}} object.
}
\description{
Constructor for lm_topic_labels objects used in this package.
}
\details{
If you call \code{as.lm_topic_labels} on an object \code{x} which already is of
the structure of an \code{lm_topic_labels} object (in particular a \code{lm_topic_labels}
object itself), the additional arguments \code{id, param, ...}
may be used to override the specific elements.
}
\examples{
# EXAMPLES TBA

}
