% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topiclabels-package.R
\docType{package}
\name{PACKAGENAME-package}
\alias{PACKAGENAME}
\alias{PACKAGENAME-package}
\title{topiclabels}
\description{
DESCRIPTION
}
\section{TITLE1}{
NA
}

\section{TITLE2}{
NA
}

\references{
REFERENCES
}
\seealso{
Useful links:
\itemize{
  \item \url{GITHUB URL}
  \item Report bugs at \url{GITHUB ISSE URL}
}

}
\author{
\strong{Maintainer}: VORNAME NACHNAME \email{mail} (\href{https://orcid.org/orcid}{ORCID})

Authors:
\itemize{
  \item VORNAME NACHNAME \email{mail} (\href{https://orcid.org/orcid}{ORCID})
  \item Jonas Rieger \email{rieger@statistik.tu-dortmund.de} (\href{https://orcid.org/0000-0002-0007-4478}{ORCID})
}

}
