% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topiclabels-package.R
\docType{package}
\name{topiclabels-package}
\alias{topiclabels}
\alias{topiclabels-package}
\title{Automated Topic Labeling with Language Models}
\description{
Leveraging (large) language models for automatic topic labeling.
The main function converts a list of top terms into a label for each topic.
Hence, it is complementary to any topic modeling package that produces a list
of top terms for each topic. While human judgement is indispensable for
topic validation (i.e., inspecting top terms and most representative documents),
automatic topic labeling can be a valuable tool for researchers in various scenarios.
}
\section{Labeling function}{

\code{\link{label_topics}}
}

\section{Constructor}{

\code{\link{as.lm_topic_labels}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PetersFritz/topiclabels}
  \item Report bugs at \url{https://github.com/PetersFritz/topiclabels/issues}
}

}
\author{
\strong{Maintainer}: Jonas Rieger \email{rieger@statistik.tu-dortmund.de} (\href{https://orcid.org/0000-0002-0007-4478}{ORCID})

Authors:
\itemize{
  \item Fritz Peters \email{fpeters3@sheffield.ac.uk} (\href{https://orcid.org/0009-0003-8471-4931}{ORCID})
  \item Andreas Fischer \email{andreasfischer1985@web.de} (\href{https://orcid.org/0009-0006-0748-6076}{ORCID})
  \item Tim Lauer \email{tl@leibniz-psychology.org} (\href{https://orcid.org/0009-0003-1625-1672}{ORCID})
  \item André Bittermann \email{abi@leibniz-psychology.org} (\href{https://orcid.org/0000-0003-2942-9831}{ORCID})
}

}
