% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_topics.R
\name{label_topics}
\alias{label_topics}
\title{TITLE TBA}
\usage{
label_topics(
  terms,
  model = "tiiuae/falcon-7b-instruct",
  params = list(),
  token,
  context = "",
  sep_terms = "; ",
  max_length_label = 5L,
  prompt_type = 1L,
  progress = TRUE
)
}
\arguments{
\item{terms}{[\code{list (k) of character}]\cr
List (each list entry represents one topic) of \code{character} vectors
containing the top terms representing the topics that are to be labeled.
If a single \code{character} vector is passed, this is interpreted as
the top terms of a single topic. If a \code{character} matrix is passed,
each column is interpreted as the top terms of a topic.}

\item{model}{[\code{character(1)}]\cr
Optional. Default is \code{tiiuae/falcon-7b-instruct}.\cr
The language model to use for labeling the topics.
The model must be accessible via the Huggingface API.}

\item{params}{[\code{list}]\cr
Optional.\cr
Model parameters to pass. Default parameters for common models are
given in the details section.}

\item{token}{[\code{character(1)}]\cr
Optional.\cr
TBA}

\item{context}{description}

\item{sep_terms}{description}

\item{max_length_label}{description}

\item{prompt_type}{description}

\item{progress}{description}
}
\value{
[\code{character(k)}] Labels for all \code{k} topics.
}
\description{
DESCRIPTION TBA
}
\details{
DETAILS TBA
}
\examples{
# EXAMPLES TBA

}
